refactor Beef.user into access, mutations, and queries directory; make ad hoc changes to other files
