+++ kousa/lib/socket_handler.ex
@@ -5,7 +5,6 @@ defmodule Kousa.SocketHandler do
   alias Kousa.RegUtils
   alias Kousa.Gen
 
-  alias Beef.Users
   alias Beef.Rooms
   alias Beef.Follows
 
@@ -111,7 +110,7 @@ defmodule Kousa.SocketHandler do
             x ->
               {user_id, tokens, user} =
                 case x do
-                  {user_id, tokens} -> {user_id, tokens, Users.get_by_id(user_id)}
+                  {user_id, tokens} -> {user_id, tokens, Beef.Access.User.get_by_id(user_id)}
                   y -> y
                 end
 
@@ -345,7 +344,7 @@ defmodule Kousa.SocketHandler do
   end
 
   def handler("speaking_change", %{"value" => value}, state) do
-    current_room_id = Users.get_current_room_id(state.user_id)
+    current_room_id = Beef.Queries.User.get_current_room_id(state.user_id)
 
     if not is_nil(current_room_id) do
       Kousa.RegUtils.lookup_and_cast(
@@ -509,7 +508,7 @@ defmodule Kousa.SocketHandler do
   end
 
   def handler("get_current_room_users", _data, state) do
-    {room_id, users} = Users.get_users_in_current_room(state.user_id)
+    {room_id, users} = Beef.Queries.User.get_users_in_current_room(state.user_id)
 
     {muteMap, autoSpeaker, activeSpeakerMap} =
       cond do
@@ -545,7 +544,7 @@ defmodule Kousa.SocketHandler do
   end
 
   def handler("ask_to_speak", _data, state) do
-    with {:ok, room_id} <- Users.tuple_get_current_room_id(state.user_id) do
+    with {:ok, room_id} <- Beef.Queries.User.tuple_get_current_room_id(state.user_id) do
       case RoomsPermission.ask_to_speak(state.user_id, room_id) do
         {:ok, %{isSpeaker: true}} ->
           Kousa.BL.Room.internal_set_speaker(state.user_id, room_id)
@@ -581,7 +580,7 @@ defmodule Kousa.SocketHandler do
   end
 
   def handler(op, data, state) do
-    with {:ok, room_id} <- Users.tuple_get_current_room_id(state.user_id),
+    with {:ok, room_id} <- Beef.Queries.User.tuple_get_current_room_id(state.user_id),
          {:ok, voice_server_id} <-
            RegUtils.lookup_and_call(Gen.RoomSession, room_id, {:get_voice_server_id}) do
       d =
